+++ app/src/main/java/com/jksalcedo/passvault/repositories/PasswordRepository.kt
@@ -67,4 +67,12 @@ class PasswordRepository(context: Context) {
     fun getEntriesByCategory(category: String): LiveData<List<PasswordEntry>> {
         return passwordDao.getEntriesByCategory(category)
     }
+
+    suspend fun getEntriesByDomain(domain: String): List<PasswordEntry> {
+        return passwordDao.getEntriesByDomain(domain)
+    }
+
+    suspend fun searchEntries(query: String): List<PasswordEntry> {
+        return passwordDao.searchEntries(query)
+    }
 }